# docs/api/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    カテゴリと難易度を自由に定義できるクイズアプリの API。

    テナント配下のエンドポイントは、触れる人によってパスが分かれる。

    - `/api/t/{slug}/admin/...` … 管理者として所属していること
    - `/api/t/{slug}/play/...` … 所属していること

    同じクイズでも管理 API は正解を含み、出題 API は含まない。
  title: quiz-app API
  version: 0.1.0
servers:
- description: アプリケーションと同じオリジン
  url: /
security:
- stubUser: []
...
paths:
...
  /api/t/{slug}/admin/quizzes:
    get:
      operationId: searchQuizzes
      parameters:
      - in: query
        name: categoryId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: difficultyId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: status
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizResponse"
          description: OK
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 利用者を特定できない
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 所属はしているが、この操作に必要な権限がない
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: テナントが存在しない、または所属していない
      summary: クイズを絞り込んで一覧
      tags:
      - クイズ（管理）
    post:
      operationId: createQuiz
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveQuizRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResponse"
          description: Created
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 利用者を特定できない
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 所属はしているが、この操作に必要な権限がない
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: テナントが存在しない、または所属していない
      summary: クイズを作成
      tags:
      - クイズ（管理）
  /api/t/{slug}/admin/quizzes/{id}:
    delete:
      operationId: deleteQuiz
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 利用者を特定できない
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 所属はしているが、この操作に必要な権限がない
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: テナントが存在しない、または所属していない
      summary: クイズを削除
      tags:
      - クイズ（管理）
    get:
      operationId: getQuiz
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResponse"
          description: OK
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 利用者を特定できない
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 所属はしているが、この操作に必要な権限がない
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: テナントが存在しない、または所属していない
      summary: クイズを 1 件取得
      tags:
      - クイズ（管理）
    put:
      operationId: updateQuiz
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveQuizRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResponse"
          description: OK
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 利用者を特定できない
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: 所属はしているが、この操作に必要な権限がない
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: テナントが存在しない、または所属していない
      summary: クイズを更新
      tags:
      - クイズ（管理）
...
components:
  schemas:
...
    ChoiceRequest:
      type: object
      properties:
        body:
          type: string
          maxLength: 500
          minLength: 0
        isCorrect:
          type: boolean
      required:
      - body
    ChoiceResponse:
      type: object
      properties:
        body:
          type: string
        id:
          type: string
          format: uuid
        isCorrect:
          type: boolean
...
    ProblemDetail:
      type: object
      additionalProperties: true
      description: RFC 9457 の Problem Details 形式のエラー応答
      properties:
        detail:
          type: string
          description: この発生に固有の説明
        instance:
          type: string
          format: uri
        status:
          type: integer
          format: int32
          description: HTTP ステータスコード
        title:
          type: string
          description: エラーの種類を表す短い説明
        type:
          type: string
          format: uri
    QuizResponse:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChoiceResponse"
        difficultyId:
          type: string
          format: uuid
        explanation:
          type: string
        id:
          type: string
          format: uuid
        question:
          type: string
        status:
          type: string
...
    SaveQuizRequest:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChoiceRequest"
        difficultyId:
          type: string
          format: uuid
        explanation:
          type: string
        question:
          type: string
          maxLength: 2000
          minLength: 0
        status:
          type: string
      required:
      - question
...
  securitySchemes:
    stubUser:
      description: Phase 1 のスタブ認証。利用者の UUID をそのまま渡す。Phase 3 で Cognito の JWT に差し替える
      in: header
      name: X-User-Id
      type: apiKey